+++ app/components/flame/card.vue
@@ -18,14 +18,20 @@ const nodePair = computed(() => toFlameNode(
     props.mode === "memory",
 ))
 const height = computed(() => Math.min(nodePair.value[1] * 20, 5000))
-const width = computed(() => container.value?.clientWidth ?? 200)
+const width = computed(() => {
+  const rect = container.value?.getBoundingClientRect()
+  if (rect) {
+    return Math.round(rect.width)
+  }
+  return 0
+})
 
 
 </script>
 
 <template>
-  <div ref="container" :class="`h-[${height}px]`" >
-    <FlameGraph :raw="nodePair[0]" :height="height" :width="width" />
+  <div ref="container" :class="`flex w-full h-[${height}px]`" style="overflow: hidden" >
+    <FlameGraph v-if="width > 0" :raw="nodePair[0]" :height="height" :width="width" />
   </div>
 </template>
