+++ app/components/flame/rect.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 
-import {constant, type ChartNode} from "~/types/flame";
+import {type ChartNode, type FlameConstants} from "~/types/flame";
 
 type Props = {
   node: ChartNode;
@@ -15,6 +15,7 @@ type Props = {
   width: number;
   x: number;
   y: number;
+  constants: FlameConstants;
 }
 
 type Emits = {
@@ -32,6 +33,7 @@ const x = ref<number>(props.x)
 const width = ref<number>(props.width)
 const height = ref<number>(props.height)
 const filter = ref<string | undefined>(undefined)
+const mounted = ref(false)
 
 onMounted(() => {
   if (x.value < 0) {
@@ -40,33 +42,34 @@ onMounted(() => {
   }
   // Fake a border by shrinking the rect slightly.
   height.value -= 1
-  width.value = Math.min(width.value, props.containerWidth)
+  width.value = Math.min(Math.round(width.value), props.containerWidth)
   filter.value = props.isDimmed ? 'brightness(115%) grayscale(50%)' : undefined
+  mounted.value = true
 })
 
 </script>
 
 <template>
-  <template>
-    <div
-        :title="disableDefaultTooltips ? undefined : tooltip != undefined ? tooltip : label"
-        :style="{
-      backgroundColor: props.backgroundColor,
-      color: props.color,
-      height: height,
-      filter: filter,
-      transform: `translate(${x}px,${props.y}px)`,
-      width: width,
-    }"
+  <div
+      v-if="mounted"
 
-        @click="emit('click')"
-        @mouseenter="emit('mouseenter', $event, props.node)"
-        @mouseleave="emit('mouseleave', $event, props.node)"
-        @mousemove="emit('mousemove', $event, props.node)"
-    >
-      &nbsp;{{ width >= constant.minWidthToDisplayText ? label : '' }}
-    </div>
-  </template>
+      :title="disableDefaultTooltips ? undefined : tooltip != undefined ? tooltip : label"
+      :style="{
+        backgroundColor: props.backgroundColor,
+        color: props.color,
+        filter: filter,
+        transform: `translate(${x}px,${props.y}px)`,
+        height: `${height}px`,
+        width: `${width}px`,
+      }"
+
+      @click="emit('click')"
+      @mouseenter="emit('mouseenter', $event, props.node)"
+      @mouseleave="emit('mouseleave', $event, props.node)"
+      @mousemove="emit('mousemove', $event, props.node)"
+  >
+    &nbsp;{{ width >= props.constants.minWidthToDisplayText ? label : '' }}
+  </div>
 </template>
 
 <style scoped>
