+++ nuxt.config.ts
@@ -18,5 +18,5 @@ export default defineNuxtConfig({
     },
     nitro: {
         preset: 'bun',
-    }
+    },
 })
